tar-sym: initial version
This utility tries to follows symlinks in a tar ball.
The current implementation handles simple redirects and recursive lookups
just fine. It can deal with file-symlinks and directory symlinks.
However, It doesn't support anything else than the "ustar  " format. It
does not have support for hardlinks, longnames or any other extensions.
Never tested the limit of the implementation, the main idea was that it
was simple and robust... And frankly, if you need anything more complicated
then you should ask, if the "tar-balls" is really the best method.

Syntax:
./tar-sym.sh find|extract file.tar path/to/possibly/symlinked/file